+++ CHANGELOG.md
@@ -11,6 +11,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fix
 
 - Dataset : Update plot to new decorator paradigm
+- Workflow : Reset variables types for block defined from user code (should not trust frontend json on variable types)
+- Workflow : If type is provided, Variables do not reset their type based on their default value anymore
+
+### Refactor
+
+- Workflow : Use AttributeType to compute variable type for ModelAttribute Blocks
 
 
 ## 0.16.2