+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## 0.16.4
+
+### Fix
+
+- Workflow : Blocks are not generating their to_dict with pointers anymore
+
+
 ## 0.16.3
 
 ### Fix